notification_service/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
python-decouple==3.8
requests==2.32.3
PyJWT==2.8.0
django-celery-results==2.5.1
celery==5.3.6